Fixed and changed player_turn, Created post_jail()
Added return statements to player_turn() so that the function runs correctly (needed the turn to end at those point and it originally was not stopping at those points.) Adjusted the code for when player is in jail so that function prints different statements. Added post_jail to take care of moving the player when the leave jail.