Adjusted/Fixed in Jail check in player_turn()
moved the self.jail_roll to the begining of the in jail if statement so that the player could roll the leave the jail at the third round in jail. When the sel.fjail_roll += 1 code was bellow the player only gets to leave after theiri 4th round in jail.