Chagned mortgage_prop() so that stops if users have nothing to mortgage
chagned mortgage_prop() so that it stops prompting the user to mortgage if they do not have any properties and prints the properties each time. Also adjusted the functiosn that call mortgage_prop() so that the player turn ends when there are no more properties to mortgage and the player still owes money.